# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.DomainSSL:
    properties:
      certificate_detail:
        description: Parsed SSL/Domain details
        type: string
      check_interval:
        description: in days
        type: integer
      created_at:
        type: string
      days_left:
        description: Calculated, not stored
        type: integer
      domain:
        type: string
      expiry:
        type: string
      expiry_threshold:
        description: Days before expiry to consider expired
        type: integer
      id:
        type: string
      issuer:
        type: string
      label:
        type: string
      resolved_ip:
        type: string
      updated_at:
        type: string
      valid_from:
        type: string
      warning_threshold:
        description: Days before expiry to warn
        type: integer
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      last_login:
        type: string
      password:
        description: Hashed password
        type: string
      role:
        description: e.g., "admin", "user"
        type: string
      status:
        description: e.g., "active", "inactive"
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Monitron Backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Monitron Backend API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return a JWT token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "Login successful", "token": "<JWT_TOKEN>", "user":
            "<USER_OBJECT>'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error": "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error": "Invalid credentials'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not generate token'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in a user
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username, email, and password
      parameters:
      - description: User object for registration
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: 'error": "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not register user'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Authentication
  /domain-ssl:
    get:
      description: Retrieve a list of all domain and SSL certificate monitoring entries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DomainSSL'
            type: array
        "500":
          description: 'error": "Could not retrieve domain/SSL entries'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all domain/SSL entries
      tags:
      - Domain & SSL
    post:
      consumes:
      - application/json
      description: Create a new domain and SSL certificate monitoring entry
      parameters:
      - description: Domain/SSL object to be created
        in: body
        name: domainSSL
        required: true
        schema:
          $ref: '#/definitions/models.DomainSSL'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DomainSSL'
        "400":
          description: 'error": "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not create domain/SSL entry'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new domain/SSL entry
      tags:
      - Domain & SSL
  /domain-ssl/{id}:
    delete:
      description: Delete a domain and SSL certificate monitoring entry by its ID
      parameters:
      - description: Domain/SSL ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: 'error": "Invalid domain/SSL ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "Domain/SSL entry not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not delete domain/SSL entry'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a domain/SSL entry
      tags:
      - Domain & SSL
    get:
      description: Retrieve a single domain and SSL certificate monitoring entry by
        its ID
      parameters:
      - description: Domain/SSL ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DomainSSL'
        "400":
          description: 'error": "Invalid domain/SSL ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "Domain/SSL entry not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not retrieve domain/SSL entry'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get domain/SSL entry by ID
      tags:
      - Domain & SSL
    put:
      consumes:
      - application/json
      description: Update details of an existing domain and SSL certificate monitoring
        entry by its ID
      parameters:
      - description: Domain/SSL ID
        in: path
        name: id
        required: true
        type: string
      - description: Domain/SSL object with updated fields
        in: body
        name: domainSSL
        required: true
        schema:
          $ref: '#/definitions/models.DomainSSL'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DomainSSL'
        "400":
          description: 'error": "Invalid domain/SSL ID" or "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "Domain/SSL entry not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not update domain/SSL entry'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an existing domain/SSL entry
      tags:
      - Domain & SSL
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"